Fix sentry additional shots and sentry flamer cause_data (#7471)
# About the pull request

This PR is a follow-up to #7406 correcting an issue where sentries
swapped the shooter and firer vars to fix #7045. It also corrects the
erroneous initialization for flamer sentries so now they will actually
attribute kills to the sentry and explain in the death message both the
sentry and owner.

# Explain why it's good for the game

Fixes additional projectiles being fired from humans that placed a
shotgun sentry; the shots are supposed to be from the sentry not
magically from the player somewhere else. Also improves the death
messages for flamer sentry kills and actually attributes kills to the
sentry.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![sentry](https://github.com/user-attachments/assets/aae7c676-216b-4c18-a7df-9781c5e83c26)

![image](https://github.com/user-attachments/assets/e7db396c-b36d-4ec8-b952-7f623a29b735)

</details>

# Changelog
:cl: Drathek
fix: Fixes additional projectiles being fired from humans that placed a
shotgun sentry instead of the sentry.
fix: Fixes cause_data for flamer sentries so now kills they make are
explained in the message and attributed to the kill counter for the
sentry.
/:cl: