More OD pragmas: Pointless positional argument, ambiguous in order, and proc argument global (#7482)
# About the pull request

This PR adds more open dream pragmas to lint for logic errors:
- PointlessPositionalArgument: A new pragma has been added for detecting
pointless positional arguments in procs, not lists
- AmbiguousInOrder: A new pragma has been added for detecting ambiguous
uses of `in` with unexpected behavior, similar to the lint in
SpacemanDMM/DreamChecker. An example is `"a" in a || "b" in b` being
parsed as `("a" in (a || "b")) in b`.
- ProcArgumentGlobal: A new pragma has been added for detecting this
BYOND bug: https://www.byond.com/forum/post/2830750
When a proc argument begins with /var/ instead of var/, it creates a
global variable instead of an argument.

# Explain why it's good for the game

More robust code.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
code: Enabled 3 more OD pragmas for linting and fixed some ambiguous
usage of In
/:cl: