Clears invalid loadout gear/restricts to actual point value. (#7477)
# About the pull request
If you're above the cap when it comes to points used in the loadout, now
the loadout item/gear will be removed.

# Explain why it's good for the game
You have a set amount of points for a reason, this hinders people
staying above the point cap after changes are made to cost or even
new/old things are added or removed.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Nanu, HarryOb
code: Adds some code so if you're above the pointcap on your loadout,
items will be removed to put you at the point cap limit.
/:cl:

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>