CL Containment Chamber De-Metagamifyer (#7427)
# About the pull request

This PR makes it so that MPs will no longer have any justification to
metagame the corporate liaison's containment cell without it being
EXTREMELY obvious.
The containment cell's divider now just looks like an ordinary wall, and
it has a custom opening and closing sprite for that purpose.

The button for the containment cell has also been moved to the back
office, as has the control panel to access it. The button has been
replaced with a button for the back office shutters, and the console has
been replaced with a TV (a thing liaisons have wanted stock standard in
their office ever since Combat Correspondent was added).

This PR also changes some minor stuff around the CL's office to make it
cleaner, and adds a WY-MRE to the cabinet, as the office had no food in
it for some reason (which has caused a few new liaisons to eat the
monkey cubes, thinking they were food cubes).

Finally, I also changed the containment cell's area name to "Storage
Room" so if an MP is snooping around the liquor cabinet, they don't see
"Containment Cell APC."

The PR contains mapping changes, of course. I know Alaymer Mapping
changes are supposed to be frozen, but I asked Nanu and he specifically
said something like this would be minor enough (and useful enough) to be
acceptable.
I tested all the changes and used mapmerge.
This PR also includes sprite additions. They shouldn't conflict with
blundir's sprite freeze, as i asked him and he said the file i modified
wasn't included in the pr. He can weigh in personally to confirm that.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

MPs metagaming the existence of the containment cell is extremely
annoying, and no one likes it. Making it much harder for them to just
say "CL open the containment cell" or "HMM, what does this button
labelled CONTAINMENT SHUTTERS on your desk DO???" is a good change.

Also the CL not having any food in his office is a tad weird, I feel
like some donks in the back makes sense
same with the TV, every CL player has wanted a TV in their office
roundstart since CC was added
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/c5b4985c-dec5-4e65-8f10-5da621745a7e)

this video doesn't properly show the desk configuration, but it shows
the animation quite well. (it was an old video)


https://github.com/user-attachments/assets/83507f49-e49d-4376-a97d-375d4f3dc62e


![image](https://github.com/user-attachments/assets/a8d896a7-f2a5-42ed-a824-a0ec37699f6a)


![image](https://github.com/user-attachments/assets/94a9d907-e25d-436b-a9d6-138b84229299)


</details>


# Changelog
:cl:
imageadd: added a new sprite for the liaison's wall divider behind his
desk
mapadd: added a WY mre to the liaison's liquor cabinet (cardboard box
full of donk pockets)
mapadd: added a tv to the liaison's desk, in place of the containment
cell console
mapadd: added a new button to the liaison's desk to replace the wall
divider button. it controls the back office shutters
maptweak: moved the liaison's shutters button and console to the back
office
maptweak: moved some stuff around in the liquor cabinet to accommodate
the MRE being placed into it
maptweak: moved some stuff around on the liaison's desk to make it look
cleaner
/:cl: