Double Facehugger Generation If Queen Is Ovi (#7490)
# About the pull request

This PR is a variation of #7440 but instead requires queen on ovi for
the cooldown reduction. I intentionally did not amend an existing
cooldown when queen mounts/dismounts.

# Explain why it's good for the game

Players have felt that there aren't enough facehuggers available, but
this variation makes it variable encouraging queen to be on ovi.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: detectivegoogle Drathek 
balance: Morphers generate huggers twice as fast if queen is on ovi
/:cl: