Pred bones can now go on armor (#7504)
# About the pull request



Yautja can now place the bone trophy ornaments on your armor instead of
only your body mesh uniform

# Explain why it's good for the game

Since you could only place the bone accessory on your mesh, it meant
that it couldn't be seen if you had armor over it. Now you can place the
bones on your armor instead and have it shown. If you still want to put
it on your mesh so that it is only seen with the armor off, you can
still do that too.


# Testing Photographs and Procedure


<details>
<summary>Screenshots & Videos</summary>
Bone on top of armor


![image](https://github.com/user-attachments/assets/c47079ae-3c25-4fc7-941d-a3ace952c191)


</details>


# Changelog
:cl: Nomoresolvalou
add: Hunters can now put their bone trophies on their armor to have it
show up there instead of on their mesh.

/:cl: