Fix Auto Oviable Areas (#7507)
# About the pull request

This PR is a follow up to #7363 and #7488 where it was discovered that
the XvX hive landmarks were being used as an area to allow ovi meaning
distress signal was getting areas not intended to be oviable. This also
makes it so the ovi restriction does not apply to XvX (all areas will
allow it).

Oviable areas were also expanded on Kutjevo, Shivas, Tijent Dam,
Chance's Claim, and Fiorina.

# Explain why it's good for the game

Ovi restrictions are intended to disallow hives next to LZs, this fixes
that. It also expands the areas oviable at the start of the game on
Kutjevo, Shivas, Trijent Dam, Chance's Claim, and Fiorina so there is
more variety in where an initial hive can be place.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![Soro](https://github.com/user-attachments/assets/ae1ec4c8-3a96-4b89-9858-1a1662da0357)

![Soro_After](https://github.com/user-attachments/assets/6a8fbcc2-97aa-43bc-9dc2-83f609c9a82b)

![Kutjevo](https://github.com/user-attachments/assets/78e3b971-9880-4322-88f5-690b99a98af9)

![Kutjevo_After](https://github.com/user-attachments/assets/6565a64c-8e83-4b67-9183-72286db3b80f)

![Shivas](https://github.com/user-attachments/assets/ec1c5cce-6490-484a-85ba-1812ce7f4dac)

![Shivas_After](https://github.com/user-attachments/assets/4d892408-914a-4bc8-9a51-d5ee439a5f00)

![Trijent](https://github.com/user-attachments/assets/530c8e40-7e6a-42f3-b80a-f1021b505011)

![Trijent_After](https://github.com/user-attachments/assets/b8a20e59-65ba-4875-b4cb-850d0c0f9073)

![Chances](https://github.com/user-attachments/assets/f70da48a-7183-464e-9a37-cd6ecbf5fa35)

![Chances_After](https://github.com/user-attachments/assets/d5f124a5-35e2-4880-a843-85c83488d90e)

![Fiorina](https://github.com/user-attachments/assets/6b58eac5-61c0-499e-a9ef-5d805d717656)

![Fiorina_After](https://github.com/user-attachments/assets/5ba8736f-290d-4b28-9338-20a17383a756)

![Lv624_After](https://github.com/user-attachments/assets/89d08415-338b-47ab-897c-f34c53832ae1)

![NV_After](https://github.com/user-attachments/assets/27ec79b5-3fc1-4d98-87f4-d408feccdc1f)

![Solaris_After](https://github.com/user-attachments/assets/24ca89c4-934a-4e61-a097-43dfae0b2030)

Testmerge with Hybrisa:

![Hybrisa](https://github.com/user-attachments/assets/5c3ba89f-7c97-428a-aa0b-519eb1680adc)

</details>


# Changelog
:cl: Drathek
balance: Fix XvX landmarks being used for roundstart oviable locations
balance: Roundstart oviable locations are expanded on Kurjevo, Shivas,
Trijent Dam, Chances's Claim, and Fiorina
fix: XvX now makes all areas oviable
/:cl: