Predator mask attachments!!! (#7494)
# About the pull request

Adds in a brand new accessory for predator masks. Simply select if you
want it on your preference menu (0 for no accessory, 1 for the
accessory) then you will receive it when you claim your equipment. The
accessory uses the same style as your mask.

# Explain why it's good for the game

More customization options for yautja is cool, and Thwomp's sprite for
the horns is cool as hell. Also, it will be incredibly easy to add in
more accessories in the future, just add in the sprites (need to be
32x64 to align correctly onmob) then change the number prompt on the
preferences menu, and you're good to go.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>
here's the bone style of the accessory with one of the masks


![image](https://github.com/user-attachments/assets/d86dbe38-9b23-4884-a681-356c2508fef6)

</details>


# Changelog
:cl: Nomoresolvalou, Thwomper
add: Adds a new predator mask accessory
ui: added a selection for mask accessories in yautja prefs
fix: fixes one speicifc yautja mask having no sprite onmob
code: infrastructure for adding in new mask accessories
imageadd: added Thwomp's cool as hell horn sprites

/:cl:

---------

Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>
Co-authored-by: harryob <me@harryob.live>