autoclimbable handrails on harm intent (#6652)
# About the pull request

you do climb when you collide with handrail, rejoice. this makes
handrails lot less of an obsticle for both huge xenos and highly rained
and agile marines. also there is autoclimbable var that can be set to
FALSE on req railing so you can not throw others out of the line


# Explain why it's good for the game

handrails can be major pain in ass for both marines and xenos, especily
as they are used near ledges so you auto climb from one side to other
but have to manualy climb back. they are also sometimes hard to spot and
can get you stuck and killed, generly getting stoped by railing that you
did not see feels like shit as both sides can shoot or slash it down in
seconds. I am gona put this as QOL but if you think it is balance then
so be it


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: handrails are autoclimbable when on harm intent, you take some harm
when doing so
balance: crusher and ravager destroy handrails when trying to pass them
similar to tables
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>