Allows bayonets to cut down fences. (#7517)
# About the pull request

This is my first PR, please take good care of me.

Adds the ability for knives to cut fences.
It now takes twice the amount of throws to break a fence.
Bayonets are slower than Wirecutters at taking down fences.

# Explain why it's good for the game

Throwing a knife to break a fence makes no sense in comparison to
cutting it with knife in-hand.

Many new players are confused by fences, expecting to take down the
fence with knife. Only ending up bashing it for a minute straight to
avail, the horrors of war.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/3f364f75-df85-44d7-9009-cfc55d39e5bc

</details>


# Changelog

:cl:
add: Added the ability for bayonets to cut down fences
balance: It now takes twice the amount of throws to bring down a fence
/:cl: