Make LZ smoke admin only (#7556)
# About the pull request

This PR is a follow up to #7363 and #7535 where the LZ miasma hazard is
now a toggleable flag during the first 3 minutes of round by admins. It
is disabled by default (as in no longer LZ miasma smoke by default).

# Explain why it's good for the game

Although the LZ smoke is successful in getting survivors out of the
landing zones, there is no longer the aspect of trying to push survivors
closer to weedable areas now that everywhere is weedable. It is awkward
for xenos if survivors back themselves up to the smoke and have no
possible flank.

The reason for making it a toggle is just because I think there can be
events that utilize it, or if staff otherwise wants to experiment with
this mechanic.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![1](https://github.com/user-attachments/assets/f0cee1fd-caed-43f2-a562-6ec3313926ed)

![2](https://github.com/user-attachments/assets/6614d334-a775-4496-b1ee-054ce6d23959)

![3](https://github.com/user-attachments/assets/a77140fc-83f1-4c0e-ad5d-b11511a733de)

</details>


# Changelog
:cl: Drathek
balance: LZ smoke is now disabled by default on distress signal
admin: Added game mode toggle to toggle LZ smoke during the first 3
minutes of the round (otherwise is locked in)
/:cl: