UPP Dropship Morana & Devana | USCM DS3 Saipan Fix (#7466)
# About the pull request

This PR adds two new UPP dropships to the game, the Morana & Devana.
They are fully functional dropships intended for use by the UPP during
admin events. They are linked to only be able to use UPP-aligned landing
zones by default, this ensures only UPP personnel can control the craft.
This can be changed by admins for events.

Both dropships are almost identical outside of some sprite differences. 

This PR also fixes the Saipan to make it useable, as previously the ship
was impossible to spawn.

Many thanks to @Zenith00000 for the sprites. 

# Explain why it's good for the game

This grants additional options for event runners to employ the use of
the UPP without needing to use a USCM dropship that has been modified
for use.


# Testing Photographs and Procedure 
(Old rear-hatch sprite)
![Screenshot 2024-11-02 00 48
27](https://github.com/user-attachments/assets/d298ea13-11ef-4250-b488-08f8c634e31d)



# Changelog
:cl: Steelpoint, Zenith
mapadd: Two UPP dropships have been added for event use, the Morana and
the Devana.
maptweak: Fixes a issue preventing the Saipan, a USCM dropship, from
being spawned by admins for events.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: harryob <me@harryob.live>