Flaregun Unique Ability Rework
Flaregun's unique action, firing an invincible flare into the sky, was not working properly nor was what I intend to work towards with flare vs xeno interaction. Gave it a niche of creating a light radius equal to a flares brightest stage for 3 minutes. Made slightly dimmer in power to not step on Mortar Flares toes, and to add atmosphere. Changed lighting subtype from Static to Movable to improve smoothness of the effect.