Code for Strata tilesets missing and restructure  (#7606)
# About the pull request
Adds the missing code needed for the strata tilesets so the tiles and
sprites are accessible to be used.

# Explain why it's good for the game
Adds the missing code needed for the strata tilesets so the tiles and
sprites are accessible to be used.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Steelpoint, Nanu
code: Adds missing strata tile-set to be available to mappers. Code
originally done by Steelpoint.
/:cl: