Buff Helmets/Gloves/Boots During HvH (#7588)
# About the pull request

A continuation of my original PR. This will buff all helmets, gloves and
boots to be comparable to the armour values of the UPP equilivants.

# Explain why it's good for the game

I've found it important from event running HvH that the USCM need
equilivant feet/hand/head protection to the UPP. This PR buffs these
items to be comparable to their UPP counterparts.

The chest area will continue to be variable, and will now be the ideal
place to try to land your shots.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: During the 'Faction Clash' (HvH) game mode. Marine Helmets,
gloves and boots will have similar armour values to their UPP
counterparts.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>