Fax Responders (#7164)
# About the pull request

This project is intended to update and overhaul the fax machines, in
regards to who can contact who and how. I have implemented some of these
changes within this PR already, but more is to come.
It also adds a dedicated role for responding to faxes, split into
subsidiary jobs for each main faction (USCM/USCM
Provost/WY/TWE/UPP/CLF/CMB)
This role is experimental, and will either be a dedicated staff role, or
potentially opened up to players. The method of this opening however is
still to be determined, whether it will act like a traditional
whitelist, or be operated differently is as of yet unknown.
There is a wiki page for the fax responder role that can be found here:
[https://cm-ss13.com/wiki/Fax_Responder](url).

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added unique identifier to all fax machines.
add: Added a new way to contact fax machines directly, titled Specific
Machine Code in the select department setting.
add: Added a Fax Responder role, broken down into subsidiary jobs for
each faction. This is currently experimental in handling, and access is
restricted.
add: Fax Machines will tell you (when selecting a faction's High
Command) if a Fax Responder is awake.
code: Makes the Whitelist Panel use the correct permissions for
Overseer.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>