Add new patches to custom loadout (#7397)
# About the pull request

Added the Solar Devils patch to custom loadout in character preferences


# Explain why it's good for the game

It's just a little and fun change. I think it can be good for character
developpement, like some marines that were before in those units.
Also in the US military, you are autorised to wear the patches of your
old units, as long as your current one is visible, so it does not break
immersion.

It's my first PR, please tell me if anything is wrong :)

# Changelog


:cl:
code: Added the Solar Devils patche to the clothing accessories for
marines

/:cl:

---------

Co-authored-by: dynastyiv <96699523+c21b4@users.noreply.github.com>
Co-authored-by: vero <73014819+vero5123@users.noreply.github.com>
Co-authored-by: harryob <me@harryob.live>