Big Red PMC crash nightmare changes + assorted fluff (#7108)
# About the pull request
Featuring these poor guys:

![replacement](https://github.com/user-attachments/assets/2187d6b0-1e83-45f2-b2ef-7e038080e484)

This PR does a few different things:

1. Adds PMC TL role to the nightmare insert.
2. Changes PMC survivor weapon pool and adds a new flavor weapon.
3. removes much of the starting spawn supplies and moves them to the
second crashed ship, which is now a guaranteed spawn.

##  New Weapon

1. WY M-39, identical to the marine version, uses ice camo.

##  PMC surv weapon pool

Originally it was:

- MP5, FP9000, Dualtube shotgun, NSG, MAR-30, custom shotgun, corp M41A

Now it is

- corp M-39, corp FP9000, Mk221, NSG, corp M41A

## The Nightmare Insert

Now there is a spot for a PMC TL, who has no extra equipment but
standard TL skills. Before, all roles would spawn with a random weapon
and full ammo belt. Now, they will spawn with a random PMC survivor
weapon with 2 spare magazines/3 spare handfulls of buckshot. Ammo and
other equipment has been placed in the second crashed ship to make up
for this loss. The supplies are similar to what the UPP crashed supply
ship has in the Trijent insert. I threw in the new WY flag because there
is no greater honor than fighting and dying beneath the banner of your
corporate overlords.

## Other minor changes

The base PMC survivor had CQC 1 and the medic and engi did not, likely
an oversight from when the medic and engi were added. Now all of them
have it. The gun lockers in eta now have the corporate M-39 instead of
the marine one. A long time ago they had the /2 ID locked version, now
harmony has been restored. I changed the faction of the survivors so
they wouldn't set off the PMC motion detector I threw in the second
ship.

# Explain why it's good for the game

Previously, the PMCs would spawn with a mish-mash of unfitting weapons
like the MAR and OP custom shotgun, which looked pretty dumb. Now, they
will spawn with corporate weapons only. Because no non-ERT corporate
M-39 existed, I made one. The team lead is a definitive leader from whom
the other members of the nightmare insert can take orders from, and also
allows for greater roleplaying opportunity if the team survives to meet
the marines. The movement of supplies from the start to the second
crashed ship gives survivors an objective to work towards early while
putting them at risk to the hive.

The logic behind the weapon replacements is that the m39 is the closest
equivalent weapon to the mp5 that makes sense, the MAR just doesn't make
any sense at all and gets dropped, and the dual-tube, which says in the
description is "used exclusively by colonial marshals" and custom
shotgun get replaced by the WY mk221. The mk221 is better than the dual
tube but worse than the custom, so it about evens out. The civ FP9000
gets replaced by the corporate one, which is more powerful only in the
most technical sense.


# Testing Photographs and Procedure
Tested by forcing the insert and spawning myself as all the new and
modified roles. The onmobs and inhands for the new weapons seem to be
fine but there are also a lot of those so I could have missed something.

<details>
<summary>Screenshots & Videos</summary>


![PMC-2](https://github.com/user-attachments/assets/434e531f-0bd4-4c75-b795-994f19a010e6)

</details>


# Changelog

:cl: BOBAMA
add: Added a new corporate version of the M-39.
add: Added new PMC TL to the Big Red nightmare insert.
add: corporate M-39, Mk221, and corporate FP9000 to PMC survivor weapon
pool to replace the MAR, FP9000, CMB shotgun, custom shotgun, and MP5.
mapadd: Changed the PMC Big Red nightmare insert to now have supplies in
the second crashed dropship near EVA.
maptweak: Changed standard M-39 gun lockers in Eta Labs to the new
corporate version.
/:cl:

---------

Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>
Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>