Adds temporary missing icon_state for gauss sentry holders (#7620)
# About the pull request

This PR is a follow up to #6363 where the temporary sprite allowing you
to undeploy the sentries somehow got lost. (It is just rappel_hatch_open
with the inside pixels transparent)

Feel free to make different sprites for the sentry holder; but this is
needed for functionality.

# Explain why it's good for the game

Allows you to undeploy gauss sentries so they can be walked over.

# Testing Photographs and Procedure

![example](https://github.com/user-attachments/assets/77acf597-c6da-444a-bf99-fa936570bc3e)

# Changelog
:cl: Drathek
imageadd: Added temporary icon for deployed state of floor sentry
/:cl: