Pill bottle emptying bugfix (#7357)
# About the pull request

Fixes an annoying bug introduced in #7265 which caused pill bottles to
run through two simultaneous emptying processes whenever used in the
hand. First trying to empty their contents, then moving to their own
pill drawing code, and trying to draw a single pill from the empty
bottle.

This PR bypasses the typical emptying system, for pill bottles and
packets, allowing a single pill to be drawn from a pill bottle in hand
with Z, but still leaves the empty verb functional, so bottles can be
emptied by selecting the option after right clicking.

# Explain why it's good for the game

While I recognize I am in the vast minority, some HM players still
prefer not to use the belt system of drawing directly from pill bottles
when treating patients.

Drawing a pill from a pill bottle held in the hand allows HMs to keep
track of which medications have been given, and in what order.

#7265 broke this system by forcing HMs to start emptying pill bottles in
the hand on a click, rather than drawing a single pill, making it
exceptionally annoying to medicate wounded without belt mode, and near
impossible to do so without a belt.

This PR allows both systems to co-exist, letting HMs draw a single pill
from bottles held in the hand if they choose to do so, and still leaving
the empty verb on pill bottles, so that their contents can be dumped
through the right click menu

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: no longer forces pill bottles to begin emptying on a click
/:cl: