More monkey logic for containers and chimpering sounds (#7636)
# About the pull request

This PR increases the odds of monkey activity when contained in an
object, adds a chance to place two chimpering sounds, and if in a locker
both a higher chance of trying to escape as well as emotes for the
locker.

# Explain why it's good for the game

Should no longer need to bap survivors for trying to hide monkeys.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/mqMKwIhK30M

</details>


# Changelog
:cl: Drathek
add: Monkeys in lockers can cause the locker to emote and have a higher
chance of escaping
soundadd: Monkeys now have a chance to use two different chimpers
(existing sfx)
/:cl: