reinforces the pipes in evac, again (#7643)
# About the pull request

I did this quite a while ago but it got overwritten in a merge conflict
so here's round two. This makes all pipes in the main evac area (from
the right of the cameras in picture below) not explode during hijack. It
also applies the same to the maintenance access corridors for the
lifeboats.

# Explain why it's good for the game

Everyone takes all the pipes out of these areas anyway, and I've made
this change previously it just got overwritten. It does make sense that
the evacuation area would have some degree of protection from disastrous
overpressure.

I could apply the same logic to CIC and reactor room, but I'm choosing
not to.


# Testing Photographs and Procedure

![image](https://github.com/user-attachments/assets/c64610dd-0a11-42f1-8d97-17524c154f57)

![image](https://github.com/user-attachments/assets/61d927f8-4fff-4108-929b-047e214e11f3)


<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Lifeboat air pipes are now reinforced (again) and no longer
explode.
/:cl: