Legacy Pred Armor Sprite Fix (#7685)
# About the pull request

Legacy pred armor did not spawn with the correct sprite because of
improper capitalization. The mask spawned in fine. Now both legacy pred
armor and the mask spawn correctly. I have tested it in game and it
works as intended, see images.

# Explain why it's good for the game

Fixes a bug with legacy gear so it can appear correctly.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

![example1](https://github.com/user-attachments/assets/ab88e000-c89e-445a-87d4-69d863a1b761)

![example2](https://github.com/user-attachments/assets/c02ed6bc-af31-4caf-b0e4-ac62c61f8420)

</details>


# Changelog
:cl:
fix: preds not spawning with correct legacy armor sprites
/:cl: