UPP Commanding Officer Vendor Addition (#7626)
# About the pull request

Adds a set of three vendors for UPP Commanding Officers (whitelisted),
similar in layout and items vended from the USCM version.

# Explain why it's good for the game

Gives more tools and better setup time for the UPP Commanders, adds
further opportunity for future item expansion as more unique items are
added.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: UPP Commanders will have access to a set of custom vendors, similar
to the vendors the USCM Commanders have access to.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>