UPP Vendor Addition | Surplus, Attachment, Weapons and Ammo (#7557)
# About the pull request

This PR adds a suite of vendors to the game for the UPP to use for
events. This includes a surplus vendor for regular soldiers to get extra
gear, and the typical vendors found in requisitions that would allow UPP
cargo personnel to vend weapons, ammo, support equipment and surplus
gear.

# Explain why it's good for the game

Further improves the OpFor experience and helps streamline cargo bay. 


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Vendor example
![Screenshot 2024-11-10 23 39
32](https://github.com/user-attachments/assets/4ae8f727-5208-4995-9886-cc7d70b553f2)

</details>


# Changelog
:cl:
add: UPP vendors have been added. Equilivants to the surplus marine
vendor and cargo vendors.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>