Some changes to broadcasting (#7662)
# About the pull request
I was writing this thing and accidentally closed my browser, so I'll be
brief.
1) You can use the camera with one hand. You still need to turn it on.
It keeps broadcasting even if dropped, but loses range. Xeno can turn
the camera off by slashing it.
2) You can unanchor TV with a wrench.
3) Broadcasting camera cant take photos anymore. But photo camera can be
used with one hand

# Explain why it's good for the game
1) More convenient.
2) If no MT is around you can still have an audience.
3) I don't think it was very convenient anyway, just use reg camera for
photos if you wanna.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
add: You can use the camera with one hand. You still need to turn it on.
It keeps broadcasting even if dropped, but loses range. Xeno can turn
the camera off by slashing it.
add: You can unanchor TV with a wrench
/:cl: