Adds ASO to Command Timelocks, and CO to MP timelocks. (#7666)
# About the pull request

As title

# Explain why it's good for the game

CO is head of Marine Law, and should count towards MP timelocks.
Likewise, ASO is third in command and should be considered a command
role.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added CO to MP timelocks, and ASO to command timelocks.
/:cl: