Fixes Typo in Paygrade Prefix for Executive Supervisor (#7715)
# About the pull request

Changes the old "Exec. Suvp." to "Exec. Spvsr."

# Explain why it's good for the game

It was both a typo and generally unclear as to what it was meant to be a
prefix for, this should clear up what the role is meant to be.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
spellcheck: Changed the old "Exec. Suvp." Paygrade Prefix to "Exec.
Spvsr."
/:cl: