Allows Ordinance Techs to take the Bad Leg trait. (#7726)
# About the pull request

title, I don't know how I missed removing OT from the list the first
time around. continuation of #4704

# Explain why it's good for the game

OT is neither a 'starter' role, or directly combat geared.

Original PR reasoning below:

"The reason that these roles specifically couldn't take this trait was
due to them being "too rookie" which I think is a bit silly to say. Even
if we take all ~~three~~ one of these to be roles only new players play
(they aren't, definitely not MT or CT) - the player still intentionally
had to go in and take the trait and therefore is assumed to accept the
traits mechanics. If they have an issue with the trait or handling it
because they are new then that is an issue that can be resolved by them
not taking detrimental traits."

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
add: OT can now take the bad leg trait.
/:cl: